refactor(transaction): rename transaction files, update imports #566
